feat: Adding changes to fix graph instance types and also adding some repo cleanup :broom:
